punch

+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/persistence/file/AsyncFileIO.java
@@ -75,7 +75,7 @@ public class AsyncFileIO extends AbstractFileIO {
     @Override public int punchHole(long position, int len) {
         int fd = 0; // TODO
 
-        return CompressionProcessor.punchHole(fd, position, len, fsBlockSize);
+        return (int)CompressionProcessor.punchHole(fd, position, len, fsBlockSize);
     }
 
     /** {@inheritDoc} */
