+++ modules/core/src/main/java/org/apache/ignite/internal/processors/compress/NativeFileSystem.java
@@ -6,5 +6,5 @@ public interface NativeFileSystem {
 
     int getFileBlockSize(Path path);
 
-    void punchHole(int fd, long off, int len);
+    void punchHole(int fd, long off, long len);
 }
